+++ libraries/CrystalStructs.h
@@ -80,7 +80,7 @@ struct Crystal_t
   TTreeFormula *FormulaTagAnalysis;
   int taggingCrystalTimingChannel;
   int taggingCrystalChannel;
-  int taggingPosition;
+  float taggingPosition;
   TCut* taggingPhotopeakCut;
   std::vector<detector_t> detectorSaturation;
   float marginWZgraph;
